requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic_settings
python-dotenv
